@@ -1,86 +1,89 @@
 {
-	"name": "react-hello-webapp",
-	"engines": {
-		"node": "16.x"
-	},
-	"version": "1.0.1",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"start": "webpack-dev-server --config webpack.dev.js --port 3000",
-		"build": "webpack --config webpack.prod.js"
-	},
-	"author": {
-		"name": "Alejandro Sanchez",
-		"url": "http://alesanchezr.com/"
-	},
-	"contributors": [
-		{
-			"name": "Alejandro Sanchez",
-			"url": "http://alesanchezr.com/"
-		},
-		{
-			"name": "Ignacio Cordoba",
-			"url": "http://github.com/nachovz"
-		}
-	],
-	"license": "ISC",
-	"devDependencies": {
-		"@babel/cli": "^7.16.0",
-		"@babel/core": "^7.16.0",
-		"@babel/plugin-proposal-class-properties": "^7.16.0",
-		"@babel/plugin-transform-runtime": "^7.16.4",
-		"@babel/preset-env": "^7.16.4",
-		"@babel/preset-react": "^7.16.0",
-		"@babel/runtime": "^7.16.3",
-		"babel-eslint": "^10.1.0",
-		"babel-loader": "^8.2.3",
-		"babel-plugin-transform-class-properties": "^6.24.1",
-		"bc-console": "0.0.2",
-		"css-loader": "^6.5.1",
-		"dotenv-webpack": "^7.0.3",
-		"envfile": "^6.17.0",
-		"error-overlay-webpack-plugin": "^1.0.0",
-		"eslint": "^8.4.0",
-		"eslint-plugin-react": "^7.27.1",
-		"eslint-webpack-plugin": "^3.1.1",
-		"file-loader": "^6.2.0",
-		"gh-pages": "^3.2.3",
-		"html-loader": "^3.0.1",
-		"html-webpack-plugin": "^5.5.0",
-		"parse-github-url": "^1.0.2",
-		"prettier": "^2.5.1",
-		"remote-origin-url": "^2.0.0",
-		"style-loader": "^3.3.1",
-		"webpack": "^5.65.0",
-		"webpack-cli": "^4.9.1",
-		"webpack-dev-server": "^4.6.0",
-		"webpack-merge": "^5.8.0"
-	},
-	"babel": {
-		"presets": [
-			"@babel/preset-env",
-			"@babel/preset-react"
-		],
-		"plugins": [
-			"@babel/plugin-proposal-class-properties",
-			[
-				"@babel/plugin-transform-runtime",
-				{
-					"regenerator": true
-				}
-			]
-		]
-	},
-	"dependencies": {
-		"@emailjs/browser": "^3.7.0",
-		"bootstrap": "^5.2.0",
-		"google-map-react": "^2.2.0",
-		"prop-types": "^15.6.1",
-		"react": "^16.8.4",
-		"react-dom": "^16.8.4",
-		"react-polyfills": "0.0.1",
-		"react-router": "^6.3.0",
-		"react-router-dom": "^6.3.0"
-	}
+  "name": "react-hello-webapp",
+  "engines": {
+    "node": "16.x"
+  },
+  "version": "1.0.1",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "start": "webpack-dev-server --config webpack.dev.js --port 3000",
+    "build": "webpack --config webpack.prod.js"
+  },
+  "author": {
+    "name": "Alejandro Sanchez",
+    "url": "http://alesanchezr.com/"
+  },
+  "contributors": [
+    {
+      "name": "Alejandro Sanchez",
+      "url": "http://alesanchezr.com/"
+    },
+    {
+      "name": "Ignacio Cordoba",
+      "url": "http://github.com/nachovz"
+    }
+  ],
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/cli": "^7.16.0",
+    "@babel/core": "^7.16.0",
+    "@babel/plugin-proposal-class-properties": "^7.16.0",
+    "@babel/plugin-transform-runtime": "^7.16.4",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/preset-react": "^7.16.0",
+    "@babel/runtime": "^7.16.3",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.3",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "bc-console": "0.0.2",
+    "css-loader": "^6.5.1",
+    "dotenv-webpack": "^7.0.3",
+    "envfile": "^6.17.0",
+    "error-overlay-webpack-plugin": "^1.0.0",
+    "eslint": "^8.4.0",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-webpack-plugin": "^3.1.1",
+    "file-loader": "^6.2.0",
+    "gh-pages": "^3.2.3",
+    "html-loader": "^3.0.1",
+    "html-webpack-plugin": "^5.5.0",
+    "parse-github-url": "^1.0.2",
+    "prettier": "^2.5.1",
+    "remote-origin-url": "^2.0.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.6.0",
+    "webpack-merge": "^5.8.0"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env",
+      "@babel/preset-react"
+    ],
+    "plugins": [
+      "@babel/plugin-proposal-class-properties",
+      [
+        "@babel/plugin-transform-runtime",
+        {
+          "regenerator": true
+        }
+      ]
+    ]
+  },
+  "dependencies": {
+    "@emailjs/browser": "^3.7.0",
+    "bootstrap": "^5.2.0",
+    "google-map-react": "^2.2.0",
+    "leaflet": "^1.9.2",
+    "prop-types": "^15.6.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-leaflet": "^4.1.0",
+    "react-polyfills": "0.0.1",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "sweetalert": "^2.1.2"
+  }
 }
