@@ -1,12 +1,15 @@
 //import react into the bundle
 import React from "react";
-import ReactDOM from "react-dom";
-
-//include your index.scss file into the bundle
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
 import "../styles/index.css";
-
-//import your own components
 import Layout from "./layout";
 
-//render your react application
-ReactDOM.render(<Layout />, document.querySelector("#app"));
+const rootElement = document.getElementById("app");
+const root = createRoot(rootElement);
+
+root.render(
+  <StrictMode>
+    <Layout />
+  </StrictMode>
+);
