@@ -1,7 +1,11 @@
 import React from "react";
 
-
 <div>
-<button onclick="location.href='aquiponemoselenlace'"  type="button" className="buttonacess" class="btn btn-warning">Incio</button>
-
-</div>
+  <button
+    onclick="location.href='aquiponemoselenlace'"
+    type="button"
+    className="buttonacess btn btn-warning"
+  >
+    Incio
+  </button>
+</div>;
