@@ -2,32 +2,34 @@ import React, { Component } from "react";
 import "../../styles/index.css";
 export const Footer = () => (
   <footer>
-    <div class="container">
-      <div class="row">
-        <div class="col-xs-12 col-md-6 text-left">
-          <h6 class="text-muted lead">KODU</h6>
-          <h6 class="text-muted">
+    <div className="container">
+      <div className="row">
+        <div className="col-xs-12 col-md-6 text-left">
+          <h6 className="d-flex justify-content-center kodutexto">KODU</h6>
+          <h6 className="text-muted d-flex justify-content-center resumen">
             Herramienta online, que actúa como punto de encuentro<br></br> entre
             las diferentes protectoras y las casas de acogida<br></br> que
             quieren compartir sus hogares con su nuevo amigo.
           </h6>
         </div>
 
-        <div class="col-xs-12 col-md-6 text-right">
-          <h6 class="text-muted lead">ENCUENTRANOS EN LAS REDES</h6>
-          <div class="redes-footer">
-            <a href="https://www.facebook.com/">
-              <img src="imagenes/facebook-2.png" />
+        <div className="col-xs-12 col-md-6 text-right">
+          <h6 className="d-flex justify-content-center kodutexto">
+            ENCUENTRANOS EN LAS REDES
+          </h6>
+          <div className="redes-footer d-flex justify-content-center h5">
+            <a href="https://www.instagram.com/" className="m-3">
+              <i className="fab fa-instagram m-auto"></i>
             </a>
-            <a href="https://twitter.com/">
-              <img src="imagenes/twitter-2.png" />
+            <a href="https://twitter.com/" className="m-3">
+              <i className="fab fa-twitter m-auto"></i>
             </a>
           </div>
         </div>
       </div>
-      <div class="row">
-        <div class="col-md-12 text-right">
-          <p class="text-muted small"> </p>
+      <div className="row">
+        <div className="col-md-12 text-right">
+          <p className="text-muted small"> </p>
         </div>
       </div>
     </div>
